requirements.txt
langchain>=0.0.267
langchain_community>=0.3.17
openai>=0.27.8
python-dotenv>=1.0.0
tiktoken>=0.4.0
pandas>=2.0.3
numpy>=1.24.3
streamlit>=1.24.1
chromadb>=0.4.6
pypdf>=3.15.1